# api/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "A server provide many function as RESTful API for processing JSON input and output the results.",
        "title": "jn",
        "contact": {
            "name": "wei840222",
            "email": "wei840222@gmail.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://github.com/wei840222/jn/blob/main/LICENSE"
        },
        "version": "0.0.1"
    },
    "paths": {
        "/invoke/js": {
            "post": {
                "description": "Run JavaScript with optional data input",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Invoke"
                ],
                "summary": "Invoke JavaScript",
                "parameters": [
                    {
                        "description": "request body",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/main.jsInvokeReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/main.jsInvokeRes"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/main.jsInvokeErrRes"
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "schema": {
                            "$ref": "#/definitions/main.jsInvokeErrRes"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "main.jsInvokeErrRes": {
            "type": "object",
            "required": [
                "error"
            ],
            "properties": {
                "error": {
                    "type": "string"
                },
                "source": {
                    "type": "string"
                },
                "stackTrace": {
                    "type": "string"
                }
            }
        },
        "main.jsInvokeReq": {
            "type": "object",
            "required": [
                "script"
            ],
            "properties": {
                "data": {
                    "type": "string",
                    "example": "test_data"
                },
                "script": {
                    "type": "string",
                    "example": "_.camelCase(data)"
                }
            }
        },
        "main.jsInvokeRes": {
            "type": "object",
            "required": [
                "result"
            ],
            "properties": {
                "result": {
                    "type": "string",
                    "example": "testData"
                }
            }
        }
    }
}